feat(app): Implement new connectivity card
closes #2555
