+++ app/src/components/RobotSettings/ConnectionCard.js
@@ -2,10 +2,8 @@
 // RobotSettings card for wifi status
 import * as React from 'react'
 import {connect} from 'react-redux'
-import {getIn} from '@thi.ng/paths'
 import find from 'lodash/find'
 
-import {getConfig} from '../../config'
 import {makeGetRobotNetworkingStatus} from '../../http-api-client'
 import {Card} from '@opentrons/components'
 import SelectNetwork from './SelectNetwork'
@@ -18,7 +16,6 @@ import type {InternetStatus, NetworkInterface} from '../../http-api-client'
 type OP = {robot: ViewableRobot}
 
 type SP = {|
-  __featureEnabled: boolean,
   internetStatus: ?InternetStatus,
   wifiNetwork: ?NetworkInterface,
   ethernetNetwork: ?NetworkInterface,
@@ -26,15 +23,10 @@ type SP = {|
 
 type Props = {...$Exact<OP>, ...SP}
 
-const __FEATURE_FLAG = 'devInternal.manageRobotConnection.newCard'
-
 export default connect(makeMapStateToProps)(ConnectionCard)
 
 const TITLE = 'Connectivity'
 function ConnectionCard (props: Props) {
-  // TODO(mc, 2018-10-15): remove feature flag
-  if (!props.__featureEnabled) return null
-
   const {robot, internetStatus, wifiNetwork, ethernetNetwork} = props
 
   return (
@@ -64,7 +56,6 @@ function makeMapStateToProps (): (State, OP) => SP {
       internetStatus,
       wifiNetwork: find(interfaces, {type: 'wifi'}),
       ethernetNetwork: find(interfaces, {type: 'ethernet'}),
-      __featureEnabled: !!getIn(getConfig(state), __FEATURE_FLAG),
     }
   }
 }
