+++ app/src/components/RobotSettings/index.js
@@ -2,12 +2,11 @@
 // robot status panel with connect button
 import * as React from 'react'
 
-import {CardContainer, CardRow, CardColumn} from '../layout'
+import {CardContainer, CardRow} from '../layout'
 import StatusCard from './StatusCard'
 import InformationCard from './InformationCard'
 import ControlsCard from './ControlsCard'
 import ConnectionCard from './ConnectionCard'
-import ConnectivityCard from './ConnectivityCard'
 import AdvancedSettingsCard from './AdvancedSettingsCard'
 import ConnectAlertModal from './ConnectAlertModal'
 import RobotUpdateModal from './RobotUpdateModal'
@@ -37,11 +36,6 @@ export default function RobotSettings (props: Props) {
       </CardRow>
       <CardRow>
         <ConnectionCard robot={robot} />
-      </CardRow>
-      <CardRow>
-        <CardColumn>
-          <ConnectivityCard robot={robot} />
-        </CardColumn>
       </CardRow>
       <CardRow>
         <AdvancedSettingsCard robot={robot} resetUrl={resetUrl} />
