+++ app/src/config/index.js
@@ -56,13 +56,6 @@ export type Config = {
   discovery: {
     candidates: string | Array<string>,
   },
-
-  // internal development flags
-  devInternal?: {
-    manageRobotConnection?: {
-      newConnectivityCard?: boolean,
-    },
-  },
 }
 
 type UpdateConfigAction = {|
